deleted package.lock.json

+++ lib/actions/user.actions.ts
@@ -36,7 +36,10 @@ export async function signInWithCredentials(
 export async function signOutUser() {
   // get current users cart and delete it so it does not persist to next user
   const currentCart = await getMyCart();
-  await prisma.cart.delete({ where: { id: currentCart?.id } });
+  if (currentCart) {
+    await prisma.cart.delete({ where: { id: currentCart?.id } });
+  }
+
   await signOut();
 }
 // Sign up a new user
